Implemeted simple api
