+++ controllers/tekmovalec.js
@@ -1,0 +1,17 @@
+import Race from "./Race.js";
+import Tekmovalec from "../data/tekmovalec.js";
+
+export default class TekmovalecController {
+    static getTekmovalec(req, res) {
+
+        return res.status(201).json({
+
+            count: Tekmovalec.length,
+
+            message: "Find one tekmovalec",
+
+            Tekmovalec: Tekmovalec
+
+        });
+    }
+}
