+++ data/tekmovalec.js
@@ -1,3 +1,19 @@
 import Race from "../controllers/Race.js";
 
-Race.MongoSetup();
+export default function Read() {
+    const client = Race.MongoSetup();
+    async function run() {
+        await client.connect();
+
+        const database = client.db("ozr");
+        const movies = database.collection("ironman");
+
+        const findResult = await movies.find({
+            name: "Hanspeter Abegg",
+          });
+        
+        findResult.forEach(console.dir);        
+        return findResult;
+    }
+    run().catch(console.dir);
+}
